Added an archive containing reapplication of various official patches
90% of these patches are reapplication of the patches
between 2016-05-03 and 2016-12-01 in
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/log/drivers/gpu/drm/rockchip

I had to try once but that's still a very stupid idea. I'll
probably drop that branch as it's completely useless to try
readapting the old branch to the new branch and, instead,
it will be better to understand how the old branch
additions work and adapt them to the new branch. It will be
FAR LESS tiring.

Signed-off-by: Myy <myy@miouyouyou.fr>