Bump to 4.10.0-rc5 and built the Mali driver as a module now.
The Mali driver is built as a module, in order to be able to update it
through an Out-Of-Tree compilation.

In the future, the module will be provided, but the kernel will not be
patched directly like it is right now. Instead, an OOT repository will
be used to compile the module against the compiled kernel.

Signed-off-by: Myy <myy@miouyouyou.fr>