The 4.10 kernel ! With Mali r16p0 drivers and the (broken) Rockchip VPU
The compilation script now uses a variable to determine which DTB
should be built. That way, it might be easier to adapt this script for
other systems, like the ASUS Tinkerboard, or the RK3288 Firefly.

A Debian package release should follow this week.

The broken VPU driver can still make it difficult to reboot. If it
clearly affect the stability of your system, do not hesitate to delete
the module. You'll be able to recompile it using the sources at :
https://github.com/Miouyouyou/rockchip-vcodec

Signed-off-by: Myy <myy@miouyouyou.fr>