The-Twelve-Myyqi : Starting the 4.12 series with the 4.12-rc1 !
Including the latest r17p0 Mali kernel drivers !

That said, almost all the patches from the previous versions got applied
without any issue... Nice to see so few internal API changes.

Anyway, the patches are actually mostly imported as-is. A "cleanup" will
be done during the evolution of this serie.

Signed-off-by: Myy <myy@miouyouyou.fr>