Yet another Bluetooth oriented commit
I love hidden linux configuration options. It's my little pleasure
after broken drivers and idiotic error messages.

There's actually a module driver for Realtek 8723 hardware named
RTLBTCOEXIST that *CANNOT* be enabled directly. You can only enable
it by enabling other drivers that depend on it. Turned out that the
staging Realtek 8723 BS drivers which used the "Bluetooth Coexist"
system didn't clearly specified that they depended on RTLBTCOEXIST so
the driver module was not compiled.

In order to enable it, I had to add other drivers that depended on the
PCI subsystem, which surprisingly was not enabled in my kernels, and
on RTLBTCOEXIST.

So this time, all the appropriate options should be enabled and we
might have a working Bluetooth driver on ASUS Tinkerboards !

If that's the case, this would resolve issue number #7 .

Signed-off-by: Myy <myy@miouyouyou.fr>