The stable branch with a 4.12.3 that I don't care about !
Seriously. I don't care about stable branch patches. If I start to care
about these patches, I'll have to maintain 4.11 stable patches, 4.12
stable patches, 4.13 stables patches and so on...

Since this repository earned me 5$ in total in 8 months, I'm not going
to sink that much time in it.

The main point of these repositories (MyyQi and RockMyy) is to make
RK3288 boards fully usable with the latest kernels, and then mainline a
maximum of patches so that people won't need any patch to use these
boards at their full capacity with the latest kernels.

Conclusion : You won't see much commits in this branch.

Signed-off-by: Myy <myy@miouyouyou.fr>